project/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
packaging
einops
transformers
setuptools
ont-bonito
causal-conv1d
pod5
optuna # added this
biopython # added this
pyslow5 # added
pysam # added